refactored Timer to separate out the Progress Circle
